requirements.txt
certifi
chardet
decorator
idna
imageio
imageio-ffmpeg
moviepy
numpy
Pillow
proglog
pytube3
requests
tqdm
typing-extensions
urllib3
